Added list_of_things_to_remove to run_from_GitHub
This should solve issues where people leave Option Explicit on when they
create scripts, so long as they capitalize it correctly. Yes I can
probably change that with a better function, but no I don't want to
right now, because this only affects scriptwriters at the moment.

Basically it uses a For...next to strip out OpEx as well as Dim-ing
case_number (which breaks a few things with functions).