Added pop-ups to warn users if they are incorrectly marked as beta
This has come up the last few days as the new document noting menu has
kicked in. Apparently many counties have a beta GlobVar file even though
the bug in the installer that caused that was fixed in January (implying
they didn't see the SIR announcement from 01.27). This will give them a
friendly reminder that they need to fix that.