Resolves #146
This function takes a parameter and determines if the user is passworded out. This can be used in conjunction with an extra do...loop around a dialog to make sure that the end user has an opportunity to get back to their dialog. 
Stearns county runs a newer version of Bluezone that causes issues when a script is running while Bluezone passwords out. This function is a workaround to allow the worker to enter the password without losing the dialog information.